requirements.txt
dlt[duckdb]>=1.22.2
apache-airflow>=2.8.0
pytest>=7.0.0